Rename exampleconfig package to exampleconfiguration
