+++ develop/tutorials/code/com.liferay.docs.exampleconfiguration/bnd.bnd
@@ -6,8 +6,9 @@ Import-Package: \
 	*
 -dsannotations: *
 Private-Package: \
-	com.liferay.docs.exampleconfig.portlet
+	com.liferay.docs.exampleconfiguration.portlet
 -includeresource:\
 	META-INF/resources=src/main/resources/META-INF/resources
 -jsp: *.jsp,*.jspf
 -plugin.jsp: com.liferay.ant.bnd.plugin.JspAnalyzerPlugin
+-metatype: *
